Mengubah data menjadi float

+++ app/src/main/java/com/example/bloo/repository/SensorRepository.kt
@@ -5,10 +5,8 @@ import androidx.lifecycle.MutableLiveData
 import com.example.bloo.model.SensorModel
 import com.google.firebase.database.*
 
-
 class SensorRepository {
     private val databaseSensor: DatabaseReference = FirebaseDatabase.getInstance().getReference("Sensor")
-
 
     private val _sensorLiveData = MutableLiveData<List<SensorModel>>()
     val sensorLiveData: LiveData<List<SensorModel>> get() = _sensorLiveData
@@ -18,10 +16,10 @@ class SensorRepository {
             override fun onDataChange(snapshot: DataSnapshot) {
                 snapshot.children.firstOrNull()?.let { data ->
                     val sensor = SensorModel(
-                        data.child("Precipitation/value").getValue(Int::class.java) ?: 0,
-                        data.child("AirTemperature/value").getValue(Int::class.java) ?: 0,
-                        data.child("SoilMoisture/value").getValue(Int::class.java) ?: 0,
-                        data.child("Humidity/value").getValue(Int::class.java) ?: 0
+                        data.child("Precipitation/value").getValue(Float::class.java) ?: 0f,
+                        data.child("AirTemperature/value").getValue(Float::class.java) ?: 0f,
+                        data.child("SoilMoisture/value").getValue(Float::class.java) ?: 0f,
+                        data.child("Humidity/value").getValue(Float::class.java) ?: 0f
                     )
                     _sensorLiveData.postValue(listOf(sensor))
                 }
